+++ docs/CHANGE_LOG.md
@@ -3,6 +3,7 @@
 ## Version 2.9.7:
 - Added "organizerEmailAddress", "url", and "title" timeline axis support.  These can also be used for the default axis setting.
 - Added a "Previous Property" button to the timeline axis header (the axis text is now also centered).
+- The tabs/content areas now use a slightly lighter background color to make them stand out a little more.
 - 
 
 <br>